cart fixed

+++ src/index/removeRecord.php
@@ -3,6 +3,8 @@
   $id_pietanza = $_REQUEST["val"];
   $sum = $_REQUEST["sum"];
 
+  $id_sess = $_SESSION['id'];
+
   $servername = "localhost";
   $username = "root";
   $password = "";
@@ -41,8 +43,24 @@
         if($updated === 0) {
           $del = "DELETE FROM pietanza_nel_ordine WHERE ID_PIETANZA='$id_pietanza'";
           mysqli_query($conn, $del);
-          $delFromOrd = "DELETE FROM ordine WHERE ID_ORDINE='$id_ordine' AND ORDINE_INVIATO=0";
-          mysqli_query($conn, $delFromOrd);
+          //Se in pietanza nel ordine non c'è più nessun id_ordine equivalente con quello in ordine allora elimino anche l'ORDINE
+          // 1. Seleziono l'ordine con id user e ordine inviato = 0
+          $takeOr = "SELECT ID_ORDINE FROM ordine WHERE ID_USER='$id_sess' AND ORDINE_INVIATO=0";
+          $execute = mysqli_query($conn, $takeOr);
+          while($ordine = $execute->fetch_array()) {
+            $getOrd = $ordine;
+          }
+          foreach ($getOrd as $key => $value) {
+            $getted = $value;
+          }
+          // 2. Controllo se in pietanza_nel_ordine c'è questo id ordine
+          $doControl = "SELECT * FROM pietanza_nel_ordine WHERE ID_ORDINE='$getted'";
+          $execControl = mysqli_query($conn, $doControl);
+          $getCount = mysqli_num_rows($execControl);
+          if($getCount === 0) {
+            $delFromOrd = "DELETE FROM ordine WHERE ID_ORDINE='$id_ordine' AND ORDINE_INVIATO=0";
+            mysqli_query($conn, $delFromOrd);
+          }
         } else {
           $ok = "UPDATE pietanza_nel_ordine SET Quantita='$updated' WHERE ID_PIETANZA='$id_pietanza'";
           mysqli_query($conn, $ok);
